various fixes in for storing character data and initial codebase modularisation
