--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,19 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredPackages">
-        <value>
-          <list size="6">
-            <item index="0" class="java.lang.String" itemvalue="numpy" />
-            <item index="1" class="java.lang.String" itemvalue="absl-py" />
-            <item index="2" class="java.lang.String" itemvalue="protobuf" />
-            <item index="3" class="java.lang.String" itemvalue="urllib3" />
-            <item index="4" class="java.lang.String" itemvalue="google-auth" />
-            <item index="5" class="java.lang.String" itemvalue="torch" />
-          </list>
-        </value>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
