+++ character.py
@@ -1,13 +1,7 @@
 import json
 import os
 from os import path
-
-
-def read(file: str):
-    return json.load(open(file))
-def write(file: str, data: dict):
-    with open(file, 'w', encoding='utf-8') as f:
-        json.dump(data, f, ensure_ascii=False, indent=4)
+from methods import write, read
 
 def create_json():
     if not path.exists('data'):
@@ -15,7 +9,7 @@ def create_json():
     if not os.path.exists('data/character.json'):
         write('data/character.json', {})
     else:
-        os.remove('data/character.json') #Added step to delete and recreate file every time game is ran
+        os.remove('data/character.json') #Added step to delete and recreate file every time game is runs
         write('data/character.json', {})
         player = read('data/character.json')
 
@@ -31,24 +25,29 @@ class Player():
         self.location = "a1"
 myPlayer = Player()
 
-
-with open("data/character.json", "r+") as f:
-    player_inventory = json.load(f)
+if os.path.exists("data/character.json"):
+        with open("data/character.json", "r+") as f:
+            player_inventory = json.load(f)
 
 def set_job(key):
     myPlayer.starting_kit = key
 
 def get_job():
+    if os.path.exists("data/character.json"):
+        with open("data/character.json", "r+") as f:
+            player_inventory = json.load(f)
     res = list(player_inventory.keys())[0]
     return res
 
     #method to get item in slot
 def get_inventory(job, key):
+    read("data/character.json")
     return player_inventory[job][key]
 
 #method to add items to inventory
 def set_inventory(job, key, value):
     player_inventory[job][key] = value
+    write("data/character.json", player_inventory)
