--- inventory.py
@@ -1,5 +1,0 @@
-import json
-from character import Player
-
-myPlayer = Player()
-
