+++ main.py
@@ -1,0 +1,5 @@
+from screens.welcome import welcome
+from character import create_json
+
+create_json()
+welcome()
