+++ methods.py
@@ -1,14 +1,8 @@
 import sys
 import time
 import os
-from os import path
-import json
 import math
-from text_input import text_input
-from map import  *
-from character import *
-
-myPlayer = Player()
+import json
 
 
 def cls():
@@ -47,6 +41,12 @@ def loading(delay: int, text: str):
         t += 1
         if t != delay: time.sleep(1)
 
+def read(file: str):
+    return json.load(open(file))
+def write(file: str, data: dict):
+    with open(file, 'w', encoding='utf-8') as f:
+        json.dump(data, f, ensure_ascii=False, indent=4)
+
 starting_kits = {
             "MEDIC": {
                 "POTIONS": 3,
@@ -71,214 +71,6 @@ starting_kits = {
                 "FOOD": 0 }
                 }
 
-player = {}
 
 loads = "Loading [▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰▰] 100%"
 
-
-def welcome():
-    welcome = """
-
-#####################################################
-       _____  __                __               
-      / ___/ / /_   ____ _ ____/ /____  _      __
-      \__ \ / __ \ / __ `// __  // __ \| | /| / /
-     ___/ // / / // /_/ // /_/ // /_/ /| |/ |/ / 
-    /____//_/ /_/ \__,_/ \__,_/ \____/ |__/|__/  
-
-    __  ___                        ____   ____   ______
-   /  |/  /____   ____   ____     / __ \ / __ \ / ____/
-  / /|_/ // __ \ / __ \ / __ \   / /_/ // /_/ // / __  
- / /  / // /_/ // /_/ // / / /  / _, _// ____// /_/ /  
-/_/  /_/ \____/ \____//_/ /_/  /_/ |_|/_/     \____/   
-                                                       
-######################################################
-           Copyright Python Task Force 2021
-
-                       [1] Play <
-                       [2] Quit <
-"""
-    type_text(welcome)
-    menu = {"1": setup_game,
-        "2": sys.exit
-       }
-    text_input("> ", menu)
-    game_menu()
-
-def setup_game():
-    type_text("""
-        _________________________________________    
-        |          WHAT IS YOUR NAME?           |
-        ‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
-                How should we call you?
-    """)
-    name = input("> ").capitalize()
-    myPlayer.name = name
-    type_text("""
-        _________________________________________    
-        |           WHAT IS YOUR JOB?           |
-        ‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
-   You can choose between Medic, Police and Tehnician.
-    """)
-    selected_kit = input("> ").upper()
-    while selected_kit not in starting_kits.keys():
-        type_text("""
-        _________________________________________    
-        |     TRY AGAIN! WHAT IS YOUR JOB?      |
-        ‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
-   You can choose between Medic, Police and Tehnician.
- Make sure you type the correct job name as written above!
-    """)
-        selected_kit = input("> ").upper()
-    else:
-        myPlayer.starting_kit = selected_kit
-    #once the player chooses the kit we will store it in data/character.json
-    player[selected_kit] = {
-        "potions": starting_kits[selected_kit]["POTIONS"],
-        "bullet": starting_kits[selected_kit]["BULLETS"],
-        "keys": starting_kits[selected_kit]["KEYS"],
-        "clues": starting_kits[selected_kit]["CLUES"],
-        "letter": starting_kits[selected_kit]["LETTERS"],
-        "food": starting_kits[selected_kit]["FOOD"],
-    }
-    write('data/character.json', player)
-    #to print their inventories
-    starting_potions = starting_kits[selected_kit]["POTIONS"]
-    starting_bullets = starting_kits[selected_kit]["BULLETS"]
-    starting_keys = starting_kits[selected_kit]["KEYS"]
-    starting_clues= starting_kits[selected_kit]["CLUES"]
-    starting_letters = starting_kits[selected_kit]["LETTERS"]
-    starting_food = starting_kits[selected_kit]["FOOD"]
-
-    type_text(f"""
-        ╔. ~༻༺~ .═══════════════════════════════╗
-                YOU ARE NOW A {myPlayer.starting_kit}
-        ╚═══════════════════════════════. ~༻༺~ .╝
-
-            You will start with {starting_potions} potions,
-                {starting_bullets} bullets and {starting_keys} keys
-                in your inventory!
-
-        The game will start in 5 seconds.....
-    """)
-
-    loading(delay=5, text=loads)
-
-
-
-##### Defined Function to call inventory
-### this will have more complex features in the future
-def display_inventory():
-    job = get_job()
-    cls()
-    inventory = f"""
-
-===============================================================
-|                                                              |
-    Name: {myPlayer.name}      Health: {myPlayer.hp}      Location: {myPlayer.location}         
-|                                                              |
-===============================================================
-|                                                              |
-|                                                              |
-|                      --- Inventory ---                       |
-|                                                              |
-|                                                              |
-===============================================================
-    [1] {get_inventory(job, "potions")} Potions                   [4] {get_inventory(job, "letter")} Letters
-    [2] {get_inventory(job, "bullet")} Bullets                   [5] {get_inventory(job, "clues")} Clues
-    [3] {get_inventory(job, "keys")} Keys                      [6] {get_inventory(job, "food")} Apples
-
-"""
-    type_text(inventory)
-    loading(5, loads)
-    game_menu()
-
-print(myPlayer.starting_kit)
-
-def help_game():
-    cls()
-    help_info = """
-
- ____  ____  ________  _____     _______   
-|_   ||   _||_   __  ||_   _|   |_   __ \  
-  | |__| |    | |_ \_|  | |       | |__) | 
-  |  __  |    |  _| _   | |   _   |  ___/  
- _| |  | |_  _| |__/ | _| |__/ | _| |_     
-|____||____||________||________||_____|    
-                                           
-                              
-Discover the mistery behind the murder of Ardit, 
-the scientist. Ardit was the brother of the Medic, 
-lover of the Police officer and best friend of the Tehnician. 
-Embark into this mistery from all three perspectives 
-and see if you can find out who murdered Ardit.
-
-        _________________________________________    
-        |          AVAILABLE COMMANDS           |
-        ‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
-  During any interaction the game will ask for your input 
-  in the from of a number which coresponds to one of the  
-             options provided by the game.
-  Make sure to type a valid answer to continue the game.
-
-        _________________________________________    
-        |        ABOUT MAP & LOCATIONS          |
-        ‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
-     Some rooms or locations are locked by default. 
-     You can use keys from your inventory to unlock 
-      them or solve the puzzle from that location.
-
-        _________________________________________    
-        |         HOW TO BEAT THE GAME?         |
-        ‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
-   Once you reach the last room and solve the puzzle 
-        (without getting killed in the process) 
-              you will finish the game.
-  """
-    type_text(help_info)
-    loading(5, loads)
-    game_menu()
-
-
-def game_menu():
-    game_options = f"""
-__________________________________________________________
-|                                                        |
-|             What do you want to do next?               |
-|   [m] Change Location | [h] Help Menu | [i] Inventory  |
-|________________________________________________________|
-    """
-    type_text(game_options)
-    menu = {
-        "m": get_direction,
-        "h": help_game,
-        "i": display_inventory
-        }
-    text_input("> ", menu)
-
-
-def get_direction():
-    cls()
-    directions = """
-_________________________________________________________
-|                                                        |
-|             Where do you want to go now?               |
-|     [s] South  | [n] North | [w] West | [e] East       |
-|________________________________________________________|
-    """
-    type_text(directions)
-    answer = input("> ")
-    while not answer.lower() in ["s","n","w","e"]:
-        print("------Please select from the available directions above------")
-        answer = input("> ")
-    if answer == "s":
-        player_movement("South")
-
-    elif answer == "n":
-        player_movement("North")
-
-    elif answer == "w":
-        player_movement("West")
-
-    elif answer == "e":
-        player_movement("East")
